Split Type.Base into separate Component classes to support Declarations
The existing Type.Base class was built on the assumption that all types use a
Definition scope (which was a fair simplification for the time). In preparation
for abstraction methods, this has been split into separate Component classes
that allow Scope.Declarations to be used. This comes with a hefty refactor, but
the end result is better and also helps with some other issues (e.g. storing the
component in IR minimizes hacky lookups for the type and also ensures tests show
inherited types in the error message, which was an annoying quirk previously).

Outside the main inheritance work, it would also be beneficial to enforce
Definition scopes at runtime (e.g. Object.type should not be abstract) and there
are a few TODO notes related to this split that leads into modularity.