Cleanup Scope casts via manual In/Out parameters
Kotlin doesn't have a way to indicate a parameter in both in/out (aka, bounded
on both sides) so this introduces a workaround by manually specifying both via
separate generic types. This makes generic uses a bit unwieldy, but removes the
need for a bunch of unnecessary casts as use-site which is worth it.