Compute Scope instance using inheritance for Component initialization
This changes Component.scope from a constructor parameter to computed field
(based off whether the component is abstract). This reduces the number of
invariants that need to be managed externally and there is currently no reason
to have this in the constructor.

Furthermore, modifiers has been changed to an optional argument (keeping it as
an argument for when additional modifiers like visibility/override are added).
In the past I've intentionally avoided default arguments to prefer specifying
them explicitly (particularly as non-test code generally shouldn't use them
anyways), but as these options grow that significantly hampers the clarity of
which values are meaningful or not. Moving forward, defaults will be more common
and the standard will be to match presence/absence in source code - e.g.
defaulting modifiers to the same values as when not specified in the source code
is a good balance as it's consistent with Rhovas itself (thus, predictable) and
test coverage is sufficient for accidental use of defaults.