Convert Variable/Function constructors to use default arguments
This continues the trend of moving towards defaults for better visibility of
meaningful values. These now work the same as the Library initializer helpers,
which requires named arguments to be used. Furthermore, parameters and returns
must always be specified since they're almost always used and tends to be better
for clarity. Overall this is an increase to real estate, but should help with
readability, consistency, and messy one-liners.

It's always possible to add secondary constructors if needed, but for now let's
see how this plays out and how much I like the named arguments approach. Similar
changes to the AST/IR at some point will feed back into this as well.