Add analysis checks for method overrides including abstract methods
This adds two categories of checks around inheritance. First, methods must
specify an `override` modifier if and only if overriding an inherited method.
Second, inherited methods must be overridden if they are abstract and the
inheriting component is not.

The implementation for this uses a messy hack with Scope generics. The component
scope now has a parent scope (`inherited`) which contains all inherited methods.
For concrete components, the child Scope.Definition will have a parent
Scope.Declaration by casting away generic safety. After validating overrides,
any abstract methods in the parent will be overriden in the child
Scope.Definition and thus cannot be resolved at runtime (currently abstract
functions are Definitions with empty bodies anyways, but this may change). This
is undoubtably hacky, but it solves the problem well and keeps inherited methods
separate while working naturally with scope resolution at runtime.

This also solves an issue where the default Struct initializers couldn't be
overriden since any new definitions would overlap with existing ones, whereas
now the default initializers are part of the inherited scope and thus properly
act as a fallback if not overridden.