Rename Modifiers.Inheritance.DEFAULT to FINAL and set constructor defaults
Quick rename for clarity and provides default values as with other environment
classes. The original intent of DEFAULT was to make it clear that it was the
"unspecified" option; but it's better to have these named properly and the
constructor default helps to convey FINAL is the default anyways. If there's a
stuation where an option can both be the default and explicitly specified, then
it may be worth revisting this (e.g., allowing `final func`).