Add README section for running Rhovas locally (latest changes)
Following an initial release of Rhovas, the plan is to keep the online editor
on the latest release (better consistency with docs, less chance of regressions,
and it has to be manually updated anyways). Therefore, testing unreleased
features (e.g. inheritance) requires running Rhovas locally with the latest
changes from master. This adds a corresponding instruction section with the
necessary commands for quick reference.