Add IR context to standard library errors (Resolves #3)
Following the introduction of a Context receiver for standard library functions,
this updates the Evaluator require/error calls used to access the current IR
(stored on the stack) for error handling. Consequently, errors will properly
identify the input range of the function invocation in diagnostics (rather than
simply showing the first line).

Note that this does not allow extracting the IR of individual arguments. This
would be nice to have following a proper push-pull context system over the
evaluator (like in the analyzer) as overall this system needs TLC anyways.