Add Context.generics for runtime generics, fix stdlib type checking (#28)
This replaces Context.arguments with Context.generics, since the only use of
arguments was to extract runtime generic types for constructing return values.
These generics are computed following the same rules as Scope resolution, but
using runtime types. Consequently, the type checking peformed when invoking
native functions should now validate generic relations as well (this is more of
an improvement as part of reified generics than an outright fix, but a key step
towards safety here).

Note that runtime times are still lacking as the Evaluator itself is using type
erasure; this only fixes the handling for the standard library.