Add Function.isResolvedBy for generic-safe resolution handling
This helper checks function resolution in a way that is generic-safe. Namely,
generic constraints must be tracked across arguments and any variants introduced
during resolution collapsed to their bounds at the end. This function is only
used in two places currently, however it is likely to be reused as part of
reified generics in the Evaluator.