Implement user-defined generics for components, fix related bugs
This allows defining components with generic types, which is already supported
internally but not exposed by the language. The implementation is
straightforward, but also fixes two bugs exposed by new test cases with generics
here (see below). Furthermore, the existing implementation for function generics
(and now component generics) does not support lower bounded varargs due to being
unable to identify which bound (lower and upper) is the generic type being
defined. There was an idea to allow apostrophes to implicitly declare generics
(as in 'T) which resolves this problem and may be added in the future. In
addition, generic definitions need to validate generic names are unique.

This also fixes two bugs. First, a few isSubtypeOf/isSupertypeOf calls within
the Type implementation did not pass the existing map of generic bindings, thus
preventing generics from being applied correctly during resolution. Second, the
check for existing property definitions used the type properties helper, which
checks the entire inheritance tree and thus does not work with inheritance
(mainly for interface properties). However, this also needs to check for setter
definitions and override restrictions in the same way as normal methods.