Validate redefined generics, refactor initialization for LinkedHashMap generics
This adds validation for redefined generic types. Additionally, component and
function declaration have been converted to use LinkedHashMaps for generics to
store the types in an easily-accessible name format. This is a can of worms in
the type system and exposes many problems related to initialization order with
generic types (namely, types need to be registered and may be accessed prior to
generic initialization completing). Library initialization has been split into
declare and define phases (similar to standard analysis) to assist with this.