Refactor Type.Reference to use generic Maps, fix generic analysis order
This makes Type.Reference consistent with Component/Function generics as in the
previous commit. During initialization, Type.GenericDelegate uses the hardcoded
names of generic types to construct the generics map correctly (as the names are
not available on the component yet). An alternative idea is to define the names
when constructing the Component, however this didn't appear to offer any
advantages and the core problem of initialization order is still present.

Additionally, this fixes the analysis order for generics to be after defining
the type for cases where the component is self-referenced as a generic bound.

This also changes Type.Generic to use a default of null (like Type.Variant) for
ease of use.