Fix generics with method/component shadowing (#27)
This fixes an issue with methods shadowing component generics. Because the
component type (with generics) is part of the `this` parameter, the method
generic is instead bound to the value of the shadowed generic in `this`. The
root cause of this problem is relying on the name for identification rather than
something like a unique id, so afterwards this restriction could be lifted.