Fix component generics being defined within static functions
Static functions are not attached to a component instance and thus shouldn't
have access to generics defined by the component. As there is no instance
parameter, this generic is not guaranteed to be used and thus should not be
defined. This also allows static functions to 'shadow' component generics, as
might be desirable within a factory method.