Remove usages of delegate inheritance
Delegation is powerful, however there are quirks related to overriding methods
implemented by delegation. This can lead to contract splicing, as overriden
methods cannot be accessed by the implementation in the delegate (hence, the
implemented interface contract is "spliced" between the delegate and the class
itself. This was an issue experienced when adding a new default method to
Function, as it defaulted to calling the delegate rather than calling the
overriden method in the defining class.

Note this is not claiming delegation itself is bad, just that this specific case
of delegation when some delegated methods are overridden can lead to this issue
of contract splicing.