Split analysis into an independent multi-phase approach
This is a major change to the analyzer that has been a long time coming. The
analyzer previously used a 3-phase approach (analyze + declare/define). Each
phase has been split into it's own class and a new Registration phase has been
added. The goal of this change is to support modules (aka multi-file analysis)
which requires multiple analyzers to advance through phases in lockstep.

After significant consideration, I believe a four-phase approach is sufficient:
 1. Registration, which defines types globally such that they can be imported
    and referenced by other files.
 2. Declaration, which declares a component type (generics and inherits) such
    that the type hierarchy is defined and subtype checks are supported.
 3. Definition, which defines a component type with associated methods following
    inheritance. This phase MUST be called in the order of the type hierarchy
    (to be implemented with modules).
 4. Analysis, which contains the bulk of analysis over executable code.

I am convinced each phase is necessary (and that phase 3 must follow DAG
traversal of the type hierarchy), however I am not convinced that these are the
only phases needed (hence, motivation for splitting these up as well).

This has been largely strewn together in the "simplest" way possible at the
expense of future abstractions, and the general approach should be reviewed
after modules are implemented and the bigger picture is a bit clearer.

Thank god for tests.