Fix regression of Type.DYNAMIC inference for list/struct literals
Using Type.isSubtypeOf for special-casing inference logic doesn't work as expected with Type.DYNAMIC, since it is a subtype of all types. This was previously done by accessing the component directly (unsafe, doesn't work with inheritance) and is now done by just checking for Type.DYNAMIC directly (not great, but acceptable for now).