Temporary fix for recursive generics causing StackOverflows
Nested generics may lead to recursive generic bindings, e.g. T = List<T>, which cause StackOverflows due to infinite substitution of the binding during typechecking. This addresses the indirect cause by short-circuiting when generic types to be compared are the same instance.

Long-term, the correct solution probably requires tearing about the current generics system and having a better understanding of scope for generics. For example, the issue about is fundamentally caused by having a generic type, say "T1", that is bound to another generic type, "T2". When these have the same name "T", the type system is unable to determine if the generic type has been bound correctly or not. This may not be necessary, but it's a complex invariant that needs to be reviewed when mixing types from different contexts (e.g. parameter generics and argument generics, or function generics and class generics) and it's unclear if this name-only approach is viable.

Generics are very, very difficult.