Add typeof/#typeof functions for manual debugging
This adds typeof (runtime) and #typeof (compile-time macro) functions for manually debugging type information. Both of these functions must be implemented as built-ins, and #typeof (as a macro) is currently hardcoded into the analyzer as general macros aren't implemented yet. This is particularly helpful for generics due to type erasure.