Support proper recursive generic bounds
Generic types can now be constructed which use recursive bounds; i.e. the bound is initialized (lazily) with a reference to the constructed generic (e.g. Equatable<T: Equatable<T : ...>>). The change here is relatively straightforward in use for creating recursive types, though userland support still needs to be added.

The largest consequence is that generics must be bound immediately in case the typecheck contains a recursive generic. This means that the bindings map is unstable and may contain incorrect bounds if the typecheck returns false. While this has not been a guarantee for how this system works, it does seem like function resolution may be using this for error messages (resulting in weird errors) and there may be other unknown usages.