Fix runtime key type of empty Map literals
Following from analyzer logic, Map literals (which are Object literals inferred into a Map type) are allowed to use any key type if the literal is empty. This allows initializing a Map variable using a literal, e.g. val map: Map<String, Integer> = {}.

Note that using a constructor is also viable, however there is an issue with generic resolution/binding that must be fixed as well for that to work as intended (and this issue applies in the same way there).