Ensure generics only bind in one direction
Generic types should only be bound when present on the left-hand side. Generics on the right-hand side are "call-site" generics are should be present in the final result, unbound.

This is part of the "Don't Cross Streams" principle, but still needs additional changes to handle generics of the same name (List<T_1> passed to List<T_2>, effectively). Additionally, calls to isSupertypeOf flip type and other, also compromising this principle.