Allow bindings to properly indicate binding direction
Generic bindings should not be bidirectional as it results in generic types with the same name but different scopes being merged together. Instead, generic bindings should apply in only one direction, or not at all if both type are from the same context. This change allows representing this binding direction for future fixes.