Implement binding refinement for non-Variant combinations
Bindings with Variant values represent upper/lower bounds on permitted types, which can be refined as additional checks are made that would restrict the bounds further.

The current algorithm works for single-inheritance, not interfaces, as it requires strict-subtyping for bounds. In other words, two interfaces A & B that are unrelated could both be upper bounds for a class because multiple interfaces can be inherited from, but typechecking will reject this as incompatible because neither type is a subtype of the other. This requires intersection types to represent.