Implement binding refinement for Variant <: Generic
Unlike other combinations, when checked in an invariant (e.g. List<*> <: List<T>) it is necessary for T to _also_ be a Variant type to allow for variance in the subtype. Since a Variant type is used for binding bounds; this leads to the bounds themselves being nested Variants. The resulting behavior appears to be correct, but there are some concerns with the nesting aspect and it reveales that the existing initial bindings are strictly incorrect.