Basic support for recursive generic bounds
Recursive generic bounds (like T: Equatable<T>) require defining the generic prior to checking the bound to avoid infinite recursion. Additionally, since the bound has it's own typechecking restrictions it needs to first be set to the _invariant_ bound, then checkend, and then updated to the _variant_ version (if applicable).

Long-term, bounds containing generics need to be checked even when looking up existing bindings as the relations between bounds are lost once the first binding is defined (since the upper bound can't represent a bindable generic type on it's own).