Remove GENERIC field on delegates for unbounded types
The GENERIC field on GenericDelegate (& friends) was intended for use with extracting generic bounds, which is now handled by the bindings helper. Unbounded types in the system must be handled carefully to ensure they are used with the correct context (namely, with Bindings set) to avoid the generic type being considered a concrete type. Removing this type fixes some problematic use cases and reduces the likelihood of future issues.