Fix refinement of inherited function generics
Inheriting a function requires rebinding the function signature with the generics of the inherited type. Previously, this process failed to refine generic bounds in the declaration properly and remove generics that become unused after binding. The impact here is small, but does affect type inference which in turn affects Lambda parameters.