Allow Tuple elements to be resolved by name
In addition to indices, Tuple elements (effectively being fields) can be given a "real" name. This is intended for use in lambda parameters, allowing a reducer to use val.acc + val.elem instead of val.0 + val.1.

However, since names exists outside of the type system (i.e. Tuple types use only the index, not name) using the name relies on the static/runtime types being consistent (which is true for relevant usages with lambda parameters). That said, the implementation of Tuple<Dynamic> relies on a hack of building a single-element Tuple with a name matching the access, e.g. tuple.10 becomes Tuple<[10: Dynamic]>, to avoid creating unnecessary elements. Giving this; opting to implement the feature now and address subtyping concerns later.