Fix missing struct inheritance while extracting property type
Struct inheritance requires analyzing fields to extract name/type. During this extraction, the struct type may be checked as part of generic bounds prior to being fully initialized, resulting in unintended failures. The only known solution to this is to initially inherit an unrefined check (which could be Struct<Dynamic>, but might as well be just Dynamic) and then recheck field types once inheritance is set the finalized struct type.