Fix super initializers constructing new instance
Previously, super initialization was hacked together by calling the super initializer and then copying initialized state to the child. This is strictly wrong for multiple reasons: 1) the parent instance has the incorrect type and is thus unable to use dynamic dispatch (this also means an instance of an abstract class can be initialized), 2) the intended guarantee of "full" initialization accessible by the parent was broken as child fields aren't set, and 3) the instance identity is not preserved which may leak unexpectedly.

The solution approach is to allowing passing an optional instance (initialized by the concrete class) to an initializer, which will be used instead of constructing a new Object instance. Ideally the initializer should be able to validate construction behavior, e.g. prevent initializing abstract classes and prevent passing an instance argument to final classes.